custom

+++ mcs/class/System.Windows.Forms/System.Windows.Forms/Message.cs
@@ -111,8 +111,9 @@ namespace System.Windows.Forms {
 		}
 
 		public override string ToString() {
-			return String.Format ("msg=0x{0:x} ({1}) hwnd=0x{2:x} wparam=0x{3:x} lparam=0x{4:x} result=0x{5:x}", msg, ((Msg) msg).ToString (), hwnd.ToInt32 (), wParam.ToInt32 (), lParam.ToInt32 (), result.ToInt32 ());
-		}
+            return String.Format("msg=0x{0:x} ({1}) hwnd=0x{2:x} wparam=0x{3:x} lparam=0x{4:x} result=0x{5:x}", msg,
+                    ((Msg) msg).ToString(), hwnd.ToInt64(), wParam.ToInt64(), lParam.ToInt64(), result.ToInt64());
+        }
 		#endregion	// Public Instance Methods
 	}
 }
